[SPOT-234][FEAT] Add MemberRemovalScheduler

+++ src/main/java/com/example/spot/web/controller/AdminController.java
@@ -4,6 +4,8 @@ import com.example.spot.api.ApiResponse;
 import com.example.spot.api.code.status.ErrorStatus;
 import com.example.spot.api.code.status.SuccessStatus;
 import com.example.spot.api.exception.GeneralException;
+import com.example.spot.scheduler.MemberRemovalScheduler;
+import com.example.spot.security.utils.SecurityUtils;
 import com.example.spot.service.admin.AdminService;
 import com.example.spot.web.dto.admin.AdminResponseDTO;
 import io.swagger.v3.oas.annotations.Operation;
@@ -29,14 +31,18 @@ public class AdminController {
 
 /* ----------------------------- 회원 정보 관리 API ------------------------------------- */
 
-    @Operation(summary = "[회원 정보 관리] 탈퇴 회원 영구 삭제 API",
+    @Operation(summary = "[회원 정보 관리] 탈퇴 회원 영구 삭제 수동 실행 API",
             description = """
             ## [회원 정보 관리] inactive 시점 이후로 30일이 지난 회원의 정보를 DB에서 영구적으로 삭제합니다.
-            * "로그인 이메일", "성명", "생년월일 정보", "진행중 스터디", \
-            "모집중 스터디", "스터디 찜 정보", "게시글", "댓글", "사진", "관심사", "관심지역"이  삭제됩니다.
+            * <탈퇴 회원 영구 삭제>는 스케줄러에 의해 오전 6시마다 실행되고 있습니다.
+            * 오류로 인하여 수동 삭제가 필요한 경우 해당 API를 호출해주시기 바랍니다.
             """)
     @DeleteMapping("/members/removal")
     public ApiResponse<AdminResponseDTO.DeletedMemberListDTO> deleteInactiveMembers() {
+        // 관리자 인증
+        if (!adminService.getIsAdmin()) {
+            throw new GeneralException(ErrorStatus._NOT_ADMIN);
+        }
         AdminResponseDTO.DeletedMemberListDTO deletedMemberListDTO = adminService.deleteInactiveMembers();
         return ApiResponse.onSuccess(SuccessStatus._MEMBER_DELETED, deletedMemberListDTO);
     }
